split to own components
