+++ bundles/mapping/shadow-plugin-3d/view/ShadowTool/Speed.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Option, Message } from 'oskari-ui';
+import { StyledSelect } from './ShadowToolStyled';
+
+export const Speed = ({ speedHandler, speed }) => {
+    const handleSpeedChange = (val) => {
+        speedHandler(val);
+    };
+
+    const speedValues = [
+        {
+            value: 'slow',
+            label: <Message messageKey={'speeds.slow'} />
+        },
+        {
+            value: 'normal',
+            label: <Message messageKey={'speeds.normal'} />
+        },
+        {
+            value: 'fast',
+            label: <Message messageKey={'speeds.fast'} />
+        }
+    ];
+    return (        
+        <StyledSelect defaultValue={speed} size='large' onChange={handleSpeedChange}>
+            {speedValues.map(speed => (
+                <Option key={speed.value} value={speed.value}>{speed.label}</Option>
+            ))}
+        </StyledSelect>
+    );
+};
+Speed.propTypes = {
+    speedHandler: PropTypes.func.isRequired,
+    speed: PropTypes.string.isRequired
+};
